Fixed getting a user's requests and other small things

+++ Scoop/Model/Path.swift
@@ -11,8 +11,8 @@ struct Path: Codable {
     var id: Int
     var startLocationPlaceId: String
     var startLocationName: String
-    var startCoords: [Int]
+    var startCoords: [Float]
     var endLocationPlaceId: String
     var endLocationName: String
-    var endCoords: [Int]
+    var endCoords: [Float]
 }
